% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridup.R
\name{grid_up}
\alias{grid_up}
\title{Display strings concisely in a figure}
\usage{
grid_up(strings, output_type = "grid", bg_colour = "gray90")
}
\arguments{
\item{strings}{A vector of strings. Note: short strings are most effective.}

\item{bg_colour}{Desired feature colour for the figure. Use html colour codes or hex codes. E.g. "Aqua" or "#00FFFF".}

\item{output_typeThe}{Desired format of the output. See documentation for examples. Either "checkerboard" or "grid".
Defaults to "grid".}
}
\value{
The figure - a gt table.
}
\description{
Display strings concisely in a figure
}
